Landing Page: - Emptied the "Industries We Serve" and "What We Offer" sections
Services Page:
- Added the "What We Offer" section below "Innovative Solution" section
